Refactored methods into utility file
